backend/requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.16.9
botocore==1.19.9
Django==3.0.7
django-storages==1.10.1
jmespath==0.10.0
pkg-resources==0.0.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.0
urllib3==1.25.11
